# backend/openapi.yaml
openapi: 3.0.0
paths:
  /auth/register:
    post:
      operationId: AuthController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
      summary: Create a new account
      tags: &ref_0
        - Auth
  /auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
      summary: Authenticate with email and password
      tags: *ref_0
  /auth/refresh:
    post:
      operationId: AuthController_refresh
      parameters: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
      summary: Refresh authentication tokens
      tags: *ref_0
  /auth/logout:
    post:
      operationId: AuthController_logout
      parameters: []
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      summary: Invalidate active refresh tokens
      tags: *ref_0
...
info:
  title: Finace API
  description: OpenAPI documentation for the Finace backend services.
  version: 1.0.0
  contact: {}
...
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    jwt:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Provide the access token obtained from the auth endpoints.
  schemas:
    RegisterDto:
      type: object
      properties:
        email:
          type: string
          example: new.user@example.com
        password:
          type: string
          example: StrongPassword1!
          minLength: 8
        name:
          type: string
          example: New User
      required:
        - email
        - password
    AuthenticatedUserDto:
      type: object
      properties:
        id:
          type: number
          example: 1
        email:
          type: string
          example: user@example.com
        name:
          type: string
          example: Jane Doe
        role:
          type: string
          enum:
            - USER
            - ADMIN
          example: USER
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - createdAt
    AuthResponseDto:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/AuthenticatedUserDto'
      required:
        - accessToken
        - refreshToken
        - user
    LoginDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
        password:
          type: string
          minLength: 8
          example: ChangeMe123!
      required:
        - email
        - password
    RefreshTokenDto:
      type: object
      properties:
        refreshToken:
          type: string
          description: Refresh token if not provided via Authorization header
          minLength: 20
...